feat(market): differentiate market-proposals: offer-proposal, offer-counter-proposal
