+++ src/market/proposal/proposal-properties.ts
